% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_synonyms.R
\name{add_synonyms}
\alias{add_synonyms}
\title{Add gene synonyms}
\usage{
add_synonyms(gene_map, syn_map)
}
\value{
\code{gene_map} \code{data.frame}
}
\description{
Add gene synonyms back into \code{gene_map}
\code{data.frame}.
}
\details{
\code{gene_map} is the output of 
\link[orthogene]{convert_orthologs}.
}
\keyword{internal}
