% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_df_type.R
\name{check_gene_df_type}
\alias{check_gene_df_type}
\title{Check gene_df}
\usage{
check_gene_df_type(gene_df, gene_col, verbose = TRUE)
}
\arguments{
\item{gene_df}{Data.frame containing the genes symbols of the input species.}

\item{gene_col}{Character string indicating either the 
input species' gene symbols are in a column with the 
input species gene symbols (\code{gene_col="<gene_col_name>"}) 
or are the row names (\code{gene_col="rownames"}) or
column names (\code{gene_col="colnames"}).}

\item{verbose}{Print messages.}
}
\value{
List of gene_df and gene_col
}
\description{
Handles gene_df regardless of whether it's a
data.frame, matrix, list, or vector
}
\keyword{internal}
