% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cortex_mrna}
\alias{cortex_mrna}
\title{Gene expression data}
\format{
An object of class \code{list} of length 2.
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/25700174/}{Publication}
}
\usage{
cortex_mrna
}
\description{
Single-cell RNA-seq gene expression matrix \code{"exp"} and 
associated per-cell metadata \code{"annot"} from mouse brain.
}
\details{
Data originally comes from Zeisel et al., 2015 (Science).
}
\examples{
\dontrun{
cortex_mrna <- ewceData::cortex_mrna()
usethis::use_data(cortex_mrna, overwrite = TRUE)
}
}
\keyword{datasets}
