% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_uninformative_genes.r
\name{drop_uninformative_genes}
\alias{drop_uninformative_genes}
\title{Drop uninformative genes}
\usage{
drop_uninformative_genes(
  exp,
  level2annot,
  DGE_method = "limma",
  min_variance_decile = NULL,
  adj_pval_thresh = 1e-05,
  drop_nonhuman_genes = FALSE,
  convert_nonhuman_genes = TRUE,
  input_species = NULL,
  as_sparse = TRUE,
  as_DelayedArray = FALSE,
  return_sce = FALSE,
  no_cores = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{exp}{Expression matrix with gene names as rownames.}

\item{level2annot}{Array of cell types, with each sequentially corresponding 
a column in the expression matrix}

\item{DGE_method}{Which method to use for the Differential Gene Expression (DGE) step.}

\item{min_variance_decile}{If \code{min_variance_decile!=NULL}, calculates the variance of the mean gene expression  across `level2annot` (i.e. cell-types),
and then removes any genes that are below \code{min_variance_decile} (on a 0-1 scale).}

\item{adj_pval_thresh}{Minimum differential expression significance
that a gene must demonstrate across \code{level2annot} (i.e. cell-types).}

\item{convert_nonhuman_genes}{Whether to convert the \code{exp} row names to human gene names.}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{as_sparse}{Convert \code{exp} to sparse matrix.}

\item{as_DelayedArray}{Convert \code{exp} to \code{DelayedArray} for scalable processing.}

\item{return_sce}{Whether to return the filtered results
as an expression matrix or a \pkg{SingleCellExperiment}.}

\item{no_cores}{Number of cores to parallelise across. 
Set to \code{NULL} to automatically optimise.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to the selected DGE method.}
}
\value{
exp Expression matrix with gene names as row names.
}
\description{
\code{drop_uninformative_genes} first removes genes 
that do not have 1:1 orthologs with humans.
}
\details{
\code{drop_uninformative_genes} then drops genes from an SCT expression matrix 
if they do not significantly vary between any celltypes.
Makes this decision based on use of an ANOVA (implemented with Limma). If 
the F-statistic for variation amongst type2 annotations
is less than a strict p-threshold, then the gene is dropped. 

A very fast alternative to DGE methods is filtering by \code{min_variance_decile}, 
which selects only genes with the top variance deciles.
}
\examples{
library(EWCE)
library(ewceData)
cortex_mrna <- cortex_mrna()
 # Use only a subset of genes to keep the example quick
cortex_mrna$exp <- cortex_mrna$exp[seq(1,300), ] 

#' ## Convert orthologs at the same time
exp2_orth <- drop_uninformative_genes(exp = cortex_mrna$exp, 
                                      level2annot = cortex_mrna$annot$level2class,
                                      drop_nonhuman_genes=TRUE, 
                                      convert_nonhuman_genes=TRUE,
                                      input_species="mouse")
}
