% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_to_hgnc.r
\name{ensembl_to_hgnc}
\alias{ensembl_to_hgnc}
\title{Gene symbols -> Ensembl IDs}
\usage{
ensembl_to_hgnc(ensembl_ids)
}
\arguments{
\item{ensembl_ids}{List of ensembl IDs.}
}
\value{
List of HGNC gene symbols.
}
\description{
Convert Ensembl IDs to HGNC gene symbols.
}
\examples{
ensembl_IDs <- c("ENSG00000128573","ENSG00000176697","ENSG00000077279","ENSG00000131095" )
gene_symbols <- ensembl_to_hgnc(ensembl_IDs)
}
