% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_mouse}
\alias{exp_mouse}
\title{Gene expression data: mouse}
\format{
sparse matrix
}
\source{
\href{https://pubmed.ncbi.nlm.nih.gov/30096314/}{Publication}
\code{
ctd <- ewceData::ctd()
exp_mouse <- as(ctd[[1]]$mean_exp, "sparseMatrix")
usethis::use_data(exp_mouse, overwrite = TRUE)
}
}
\usage{
data("exp_mouse")
}
\description{
Mean pseudobulk single-cell RNA-seq gene expression matrix.
}
\details{
Data originally comes from Zeisel et al., 2018 (Cell).
}
\keyword{datasets}
