% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cache_save_path.R
\name{get_cache_save_path}
\alias{get_cache_save_path}
\title{Get cache save path}
\usage{
get_cache_save_path(
  fn,
  species,
  method,
  suffix = ".csv.gz",
  save_dir = cache_dir()
)
}
\arguments{
\item{fn}{Function name.}

\item{species}{input species name.}

\item{method}{Method (gprofiler2, homologene, or babelgene).}

\item{suffix}{Cache file suffix.}

\item{save_dir}{Cache file directory.}
}
\value{
Path string
}
\description{
Create a save path for a cached file.
}
\keyword{internal}
