% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orgdb_genomeinfodbdata.R
\name{get_orgdb_genomeinfodbdata}
\alias{get_orgdb_genomeinfodbdata}
\title{Import organism database: GenomeInfoDbData}
\source{
\href{https://github.com/Bioconductor/GenomeInfoDbData/blob/master/data/specData.rda}{
GenomeInfoDbData GitHub}
}
\usage{
get_orgdb_genomeinfodbdata(verbose = TRUE)
}
\value{
Organisms \code{data.table}
}
\description{
Import and format organism ID table from
\pkg{GenomeInfoDbData} to 
be comparable to \code{get_orgdb_gprofiler}.
}
\keyword{internal}
