% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtree_plot.R
\name{ggtree_plot}
\alias{ggtree_plot}
\title{Plot a phylogenetic tree}
\usage{
ggtree_plot(
  tr,
  d,
  scaling_factor = 1,
  clades = NULL,
  clades_palette = NULL,
  reference_species = NULL,
  verbose = TRUE
)
}
\arguments{
\item{tr}{Tree.}

\item{d}{Metadata}

\item{scaling_factor}{How much to scale y-axis parameters (e.g. offset) by.}

\item{clades}{Clades metadata.}

\item{clades_palette}{Palette to color highlighted \code{clades} with.}

\item{verbose}{Print messages.}
}
\description{
Plot a phylogenetic tree with ggtree and metadata from
 \link[orthogene]{report_orthologs}.
}
\keyword{internal}
