% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gprofiler_namespace}
\alias{gprofiler_namespace}
\title{\link[gprofiler2]{gconvert} namespaces}
\format{
\code{data.frame}
}
\source{
\href{https://biit.cs.ut.ee/gprofiler/page/namespaces-list}{
gProfiler site}

\code{
 #### Manually-prepared CSV ####
 path <- "inst/extdata/gprofiler_namespace.csv.gz"
 gprofiler_namespace <- data.table::fread(path)  
}
}
\description{
Available namespaces used by \link[gprofiler2]{gconvert}.
}
