% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gprofiler_orgs}
\alias{gprofiler_orgs}
\title{Reference organisms}
\format{
\code{data.frame}
}
\source{
\href{https://biit.cs.ut.ee/gprofiler/gost}{gProfiler site}

\code{
# NOTE!: Must run usethis::use_data for all internal data at once.
# otherwise, the prior internal data will be overwritten.
#### Internal data 1: gprofiler_namespace ####
 #### Manually-prepared CSV ####
 path <- "~/Downloads/gconvert_namespace.csv"
 gprofiler_namespace <- data.table::fread(path)  
 #### Internal data 2: gprofiler_orgs
 gprofiler_orgs <- orthogene:::get_orgdb_gprofiler(use_local=FALSE)
 #### Save ####
 usethis::use_data(gprofiler_orgs,gprofiler_namespace,
  overwrite = TRUE, internal=TRUE) 
}
}
\usage{
gprofiler_orgs
}
\description{
Organism for which gene references are available via
\href{https://biit.cs.ut.ee/gprofiler/gost}{gProfiler}
\href{https://biit.cs.ut.ee/gprofiler/api/util/organisms_list}{API}.

Used as a backup if API is not available.
}
\keyword{datasets}
