% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgnc_to_ensembl.R
\name{hgnc_to_ensembl}
\alias{hgnc_to_ensembl}
\title{Ensembl IDs -> Gene symbols}
\usage{
hgnc_to_ensembl(gene_symbols)
}
\arguments{
\item{gene_symbols}{List of HGNC gene symbols.}
}
\value{
List of ensembl IDs.
}
\description{
Convert HGNC gene symbols to Ensembl IDs.
}
\examples{
gene_symbols <- c("FOXP2","BDNF","DCX","GFAP")
ensembl_IDs <- hgnc_to_ensembl(gene_symbols)
}
