% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_species_plot.R
\name{infer_species_plot}
\alias{infer_species_plot}
\title{infer_species_plot}
\usage{
infer_species_plot(matches, show_plot = TRUE)
}
\value{
ggplot object.
}
\description{
Plot results from \link[orthogene]{infer_species}.
}
\keyword{internal}
