% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_dictionary.R
\name{invert_dictionary}
\alias{invert_dictionary}
\title{Invert dictionary}
\usage{
invert_dictionary(dict)
}
\value{
Named list
}
\description{
Switch the names/items in a named list.
}
\keyword{internal}
