% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_genes.R
\name{map_genes}
\alias{map_genes}
\title{Map genes}
\usage{
map_genes(
  genes,
  species = "hsapiens",
  target = "ENSG",
  mthreshold = Inf,
  filter_na = FALSE,
  numeric_ns = "",
  verbose = TRUE
)
}
\arguments{
\item{genes}{Gene list.}

\item{species}{Species to map against.}

\item{target}{target namespace.}

\item{mthreshold}{maximum number of results per initial alias to show. Shows all by default.}

\item{filter_na}{logical indicating whether to filter out results without a
corresponding target.}

\item{numeric_ns}{namespace to use for fully numeric IDs (\href{https://biit.cs.ut.ee/gprofiler/page/namespaces-list}{list of available namespaces}).}

\item{verbose}{Print messages.}
}
\value{
Table with standardised genes.
}
\description{
Input a list of genes in any format 
(e.g. HGNC symbols, ENSEMBL IDs, UCSC) and return
a table with standardised gene symbols
 (the "names" column).
}
\details{
Uses \code{gprofiler2::gconvert()}. 
The exact contents of the output table will depend on 
\code{target} parameter. 
See \code{?gprofiler2::gconvert} for more details.
}
\examples{
data("exp_mouse")
mapped_genes <- map_genes(genes=rownames(exp_mouse)[seq(1,100)],
                          species="mouse")
}
