% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_matrix.R
\name{map_matrix}
\alias{map_matrix}
\title{Convert genes from one species to another}
\usage{
map_matrix(
  obj,
  input_species,
  output_species = "human",
  drop_nonorths = TRUE,
  one_to_one_only = TRUE,
  convert_nonhuman_genes = TRUE,
  as_sparse = FALSE,
  sort_rows = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{Data object to be converted. Can be a matrix, data.frame, data.table.}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{output_species}{Name of the output species (e.g. "human").}

\item{drop_nonorths}{Drop genes that don't have an ortholog in the \code{output_species}.}

\item{one_to_one_only}{Drop genes that don't have a 1:1 mapping between input species and human
(i.e. drop genes with multiple human orthologs).}

\item{as_sparse}{If \code{obj} is a matrix, it can be converted to a sparse matrix.}

\item{sort_rows}{Sort rows alphanumerically.}

\item{verbose}{Print messages.}
}
\description{
Maps genes across species and then converts them to the \code{ouput_species}.
}
\examples{
data("cortex_mrna")
exp_human <- convert_orthologs(obj=cortex_mrna$exp, input_species="mouse")
}
