% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orthologs_babelgene.R
\name{map_orthologs_babelgene}
\alias{map_orthologs_babelgene}
\title{Map orthologs: babelgene}
\source{
\href{https://cran.r-project.org/web/packages/babelgene/vignettes/babelgene-intro.html}{
babelgene tutorial}
}
\usage{
map_orthologs_babelgene(
  genes,
  input_species,
  output_species = "human",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{genes}{Gene list.}

\item{input_species}{Name of the input species (e.g., "mouse","fly").
Use \link[orthogene]{map_species} to return a full list
of available species.}

\item{output_species}{Name of the output species (e.g. "human","chicken").
Use \link[orthogene]{map_species} to return a full list
 of available species.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed
to \link[babelgene]{orthologs}.}
}
\value{
Ortholog map \code{data.frame}
}
\description{
Map orthologs from one species to another
using \link[babelgene]{orthologs}.
}
\keyword{internal}
