% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orthologs_gorth.R
\name{map_orthologs_gorth}
\alias{map_orthologs_gorth}
\title{Map orthologs: gorth}
\usage{
map_orthologs_gorth(
  genes,
  input_species,
  output_species = "human",
  filter_na = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{genes}{Gene list.}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{output_species}{Name of the output species (e.g. "human").}

\item{filter_na}{logical indicating whether to filter out results without a
corresponding target name.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to
\code{gprofiler2::gorth()}.}
}
\value{
Ortholog map \code{data.frame}
}
\description{
Map orthologs from one species to another
using \code{gprofiler2::gorth()}.
}
