% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orthologs_homologene.R
\name{map_orthologs_homologene}
\alias{map_orthologs_homologene}
\title{Map orthologs: homologene}
\usage{
map_orthologs_homologene(
  genes,
  input_species,
  output_species = "human",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{genes}{Gene list.}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{output_species}{Name of the output species (e.g. "human").}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed
to \link[homologene]{homologene}.}
}
\value{
Ortholog map \code{data.frame}
}
\description{
Map orthologs from one species to another 
using \link[homologene]{homologene}.
}
\keyword{internal}
