% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{orthogene-package}
\alias{orthogene}
\alias{orthogene-package}
\title{Interspecies gene mapping}
\source{
\itemize{
\item{\href{https://github.com/neurogenomics/orthogene}{GitHub} : }{
Source code and Issues submission.}
\item{\href{https://bschilder.github.io/BMSchilder/}{Author Site} : }{
\code{orthogene} was created by Brian M. Schilder.}  
}
}
\description{
\pkg{orthogene} is an R package for easy mapping of 
orthologous genes across hundreds of species.
}
\details{
It pulls up-to-date interspecies gene ortholog mappings across 700+ organisms.\cr
It also provides various utility functions to map common objects 
(e.g. data.frames, gene expression matrices, lists) onto 
1:1 gene orthologs from any other species.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neurogenomics/orthogene}
  \item Report bugs at \url{https://github.com/neurogenomics/orthogene/issues}
}

}
\author{
\strong{Maintainer}: Brian Schilder \email{brian_schilder@alumni.brown.edu} (\href{https://orcid.org/0000-0001-5949-2191}{ORCID})

}
