% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_report.R
\name{ortholog_report}
\alias{ortholog_report}
\title{Identify the number of orthologous genes between two species}
\usage{
ortholog_report(
  target_species = "mouse",
  reference_species = "human",
  drop_nonorths = TRUE,
  one_to_one_only = TRUE,
  round_digits = 1,
  return_report = TRUE
)
}
\arguments{
\item{target_species}{Target species.}

\item{reference_species}{Reference species.}

\item{drop_nonorths}{Drop genes that don't have an ortholog in the \code{output_species}.}

\item{one_to_one_only}{Drop genes that don't have a 1:1 mapping between input species and human
(i.e. drop genes with multiple human orthologs).}

\item{round_digits}{Number of digits to round to when printing percentages.}

\item{return_report}{Return just the ortholog mapping between two species (\code{FALSE}) 
or return both the ortholog mapping as well a dataframe of the report statistics (\code{TRUE}).}
}
\description{
\code{ewceData::mouse_to_human_homologs()} provides 15,604 1:1 mouse:human orthologs.
\code{homologene}-based functions like \code{EWCE::convert_orthologs} 
provide an improved 16,766 1:1 mouse:human orthologs.
}
\examples{
orth.fly <- ortholog_report(target_species = "fly",
                            reference_species="human")
orth.zeb <- ortholog_report(target_species = "zebrafish", 
                                  reference_species="human")
orth.mus <- ortholog_report(target_species = "mouse", 
                                  reference_species="human") 
}
