% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benchmark_scatter.R
\name{plot_benchmark_scatter}
\alias{plot_benchmark_scatter}
\title{Plot benchmark: scatter}
\usage{
plot_benchmark_scatter(
  bench_res,
  remove_failed_times = FALSE,
  show_plot = TRUE
)
}
\arguments{
\item{bench_res}{Results from}

\item{remove_failed_times}{In instances where 
no genes were returned, set time to \code{NA}.}

\item{show_plot}{Print plot.}
}
\value{
ggplot object
}
\description{
Plot run time vs. # genes returned across species and 
function tests.
}
\keyword{internal}
