% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_orthotree.R
\name{plot_orthotree}
\alias{plot_orthotree}
\title{Create a phylogenetic tree of shared orthologs}
\source{
\href{https://yulab-smu.top/treedata-book/chapter8.html#phylopic}{
ggtree tutorial}
}
\usage{
plot_orthotree(
  tree = NULL,
  orth_report = NULL,
  species = NULL,
  method = c("homologene", "gprofiler", "babelgene"),
  reference_species = "human",
  clades = list(Primates = c("Homo sapiens", "Macaca mulatta"), Eutherians =
    c("Homo sapiens", "Mus musculus", "Bos taurus"), Mammals = c("Homo sapiens",
    "Mus musculus", "Bos taurus", "Ornithorhynchus anatinus", "Monodelphis domestica"),
    Tetrapods = c("Homo sapiens", "Mus musculus", "Gallus gallus", "Anolis carolinensis",
    "Xenopus tropicalis"), Vertebrates = c("Homo sapiens", "Mus musculus",
    "Gallus gallus", "Anolis carolinensis", "Xenopus tropicalis", "Danio rerio")),
  scaling_factor = 1,
  show_plot = TRUE,
  save_paths = c(tempfile(fileext = ".ggtree.pdf"), tempfile(fileext = ".ggtree.png")),
  width = 10,
  height = 10,
  mc.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{tree}{A phylogenetic tree of class \link[ape]{phylo}. If no tree 
is provided (\code{NULL}) a 100-way multiz tree will be imported from  
\href{http://hgdownload.soe.ucsc.edu/goldenPath/hg38/multiz100way}{
UCSC Genome Browser}.}

\item{orth_report}{An ortholog report from one or more species 
generated by \link[orthogene]{report_orthologs}.}

\item{species}{Species to include in the final plot.
If \code{NULL}, then all species from the given database (\code{method}) 
will be included (via \link[orthogene]{map_species}), 
so long as they also exist in the \code{tree}.}

\item{method}{R package to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives consensus scores for each gene mapping based on a
 several different data sources.}
}}

\item{reference_species}{Reference species.}

\item{clades}{A named list of clades each containing list fo species
to define the respective clade using \link[ggtree]{MRCA}.}

\item{scaling_factor}{How much to scale y-axis parameters (e.g. offset) by.}

\item{show_plot}{Whether to print the final tree plot.}

\item{save_paths}{Paths to save plot to.}

\item{width}{Saved plot width.}

\item{height}{Saved plot height.}

\item{mc.cores}{Number of cores to parallelise different steps with.}

\item{verbose}{Print messages.}
}
\value{
A list containing:  
\itemize{
\item{plot : }{Annotated ggtree object.}
\item{tree : }{The pruned, standardised phylogenetic tree used in the plot.}
\item{orth_report : }{Ortholog reports for each \code{species} against
 the \code{reference_species}}.
\item{metadata : }{Metadata used in the plot,
 including silhouette PNG ids from \href{http://phylopic.org/}{phylopic}.}
\item{clades : }{Metadata used for highlighting clades.}
\item{method : }{\code{method} used.}
\item{reference_species : }{\code{reference_species} used.}
\item{save_paths : }{\code{save_paths} to plot.}
}
}
\description{
Automatically creates a phylogenetic tree plot annotated with metadata
describing how many orthologous genes each species shares with the
 \code{reference_species} ("human" by default).
}
\examples{
orthotree <- orthogene::plot_orthotree(species = c("human","monkey","mouse"),
                                       scaling_factor = .3)
}
