% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_orthologs.R
\name{report_orthologs}
\alias{report_orthologs}
\title{Report orthologs}
\usage{
report_orthologs(
  target_species = "mouse",
  reference_species = "human",
  standardise_genes = FALSE,
  method_all_genes = c("gprofiler", "homologene"),
  method_convert_orthologs = c("gprofiler", "homologene"),
  drop_nonorths = TRUE,
  non121_strategy = "drop_both_species",
  round_digits = 2,
  return_report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{target_species}{Target species.}

\item{reference_species}{Reference species.}

\item{standardise_genes}{If \code{TRUE} genes will first be standardised 
to HGNC symbols using \link[gprofiler2]{gconvert}. 
This will add a new column to \code{gene_df} 
called "input_gene_standard".}

\item{method_all_genes}{R package to to use in 
\link[orthogene]{all_genes} step: 
\code{"gprofiler"} (slower but more species and genes) or 
\code{"homologene"} (faster but fewer species and genes).}

\item{method_convert_orthologs}{R package to to use in 
\link[orthogene]{convert_orthologs} step: 
\code{"gprofiler"} (slower but more species and genes) or 
\code{"homologene"} (faster but fewer species and genes).}

\item{drop_nonorths}{Drop genes that don't have an ortholog 
in the \code{output_species}.}

\item{non121_strategy}{How to handle genes that don't have 
1:1 mappings between \code{input_species}:\code{output_species}.
Options include: 

\itemize{
\item{\code{"drop_both_species" or "dbs" or 1} : }{Drop genes that have duplicate 
mappings in either the \code{input_species} or \code{output_species} 
(\emph{Default}).} 
\item{\code{"drop_input_species" or "dis" or 2} : }{Drop genes that have duplicate 
mappings in the \code{input_species}.} 
\item{\code{"drop_output_species" or "dos" or 3} : }{Drop genes that have duplicate
mappings in the \code{output_species}.}
\item{\code{"keep_both_species" or "kbs" or 4} : }{Keep all genes regardless of whether
they have duplicate mappings in either species.}
}}

\item{round_digits}{Number of digits to round to when printing percentages.}

\item{return_report}{Return just the ortholog mapping 
between two species (\code{FALSE}) or return both the
 ortholog mapping as well a \code{data.frame} 
 of the report statistics (\code{TRUE}).}

\item{verbose}{Print messages.}
}
\value{
List of ortholog report statistics
}
\description{
Identify the number of orthologous genes between two species.
}
\examples{
orth_fly <- report_orthologs(target_species="fly",
                             reference_species="human")
}
