% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_gprofiler.R
\name{set_gprofiler}
\alias{set_gprofiler}
\title{Set gprofiler}
\usage{
set_gprofiler(url = "http://biit.cs.ut.ee/gprofiler_beta")
}
\arguments{
\item{url}{the base URL.}
}
\value{
Null
}
\description{
Set the default URL for gprofiler API queries.
\describe{
\item{default}{http://biit.cs.ut.ee/gprofiler}
\item{bea|}{http://biit.cs.ut.ee/gprofiler_beta}
}
}
\keyword{internal}
