% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_id_dict.R
\name{taxa_id_dict}
\alias{taxa_id_dict}
\title{Taxa ID dictionary}
\usage{
taxa_id_dict(
  species = c("human", "chimp", "monkey", "mouse", "rat", "dog", "cow", "chicken",
    "zebrafish", "frog", "fly", "worm", "rice")
)
}
\arguments{
\item{species}{Species to get dictionary for. 
Can supply either Latin names (e.g. "Homo sapiens") or 
common names (e.g, "human").}
}
\value{
Named list of taxa IDs to organism names.
}
\description{
Dictionary of NCBI taxonomy IDs mapped to 
Latin and common names of 20+ organisms.
}
\examples{
dict <- taxa_id_dict()
full_dict <- taxa_id_dict(species=NULL) 
print(full_dict[c("chimp","mouse")])
}
