% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_rows.R
\name{aggregate_rows}
\alias{aggregate_rows}
\title{Aggregate rows of matrix}
\usage{
aggregate_rows(
  X,
  groupings,
  FUN = "sum",
  method = c("monocle3", "stats", "delayedarray"),
  as_sparse = TRUE,
  as_DelayedArray = TRUE,
  dropNA = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{X}{Input matrix.}

\item{groupings}{Gene groups of the same length as \code{nrow(X)}.}

\item{FUN}{Aggregation function (\emph{DEFAULT:} \code{"sum"}).}

\item{method}{Aggregation method.}

\item{as_sparse}{Convert aggregated matrix to sparse matrix.}

\item{as_DelayedArray}{Convert aggregated matrix to
\link[DelayedArray]{DelayedArray}.}

\item{dropNA}{Drop genes assigned to \code{NA} in \code{groupings}.}

\item{verbose}{Print messages.}
}
\value{
Aggregated matrix
}
\description{
Aggregate rows of a matrix using a grouping vector.
}
\keyword{internal}
