% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_genes.R
\name{all_genes}
\alias{all_genes}
\title{Get all genes}
\usage{
all_genes(species, verbose = TRUE)
}
\arguments{
\item{species}{Species to get genes for. 
Will first be standardised with \code{map_species}.}

\item{verbose}{Print messages.}
}
\value{
Gene table
}
\description{
Return all known genes from a given species.
}
\details{
References \code{homologene::homologeneData}.
}
\examples{
genome_mouse <- all_genes(species="mouse")
genome_human <- all_genes(species="human")
}
