% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gprofiler_orgs}
\alias{gprofiler_orgs}
\title{Reference organisms}
\format{
\code{data.frame}
\code{
 gprofiler_orgs <- jsonlite::fromJSON('https://biit.cs.ut.ee/gprofiler/api/util/organisms_list')
 gprofiler_orgs <-  dplyr::arrange(gprofiler_orgs, scientific_name)
 usethis::use_data(gprofiler_orgs, overwrite = TRUE, internal=TRUE)
}
}
\source{
\href{https://biit.cs.ut.ee/gprofiler/gost}{gProfiler site}
}
\usage{
gprofiler_orgs
}
\description{
Organism for which gene references are available via  
\href{https://biit.cs.ut.ee/gprofiler/gost}{gProfiler}  
\href{https://biit.cs.ut.ee/gprofiler/api/util/organisms_list}{API}. 

Used as a backup if API is not available.
}
\keyword{datasets}
