% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orthologs.R
\name{map_orthologs}
\alias{map_orthologs}
\title{Map orthologs}
\usage{
map_orthologs(
  genes,
  standardise_genes = FALSE,
  input_species,
  output_species = "human",
  method = c("gprofiler", "homologene"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{genes}{can be a mixture of any format
(HGNC symbols, ENSEMBL, UCSC, etc.)
and will be automatically converted to
 standardised HGNC symbol format.}

\item{standardise_genes}{If \code{TRUE} genes will first be standardised 
to HGNC symbols using \link[gprofiler2]{gconvert}. 
This will add a new column to \code{gene_df} 
called "input_gene_standard".}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{output_species}{Name of the output species (e.g. "human").}

\item{method}{R package to to use for gene mapping: 
\code{"gprofiler"} (slower but more species and genes) or 
\code{"homologene"} (faster but fewer species and genes).}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to 
\link[gprofiler2]{gconvert} or \link[homologene]{homologene}.}
}
\value{
Ortholog map \code{data.frame} with at
 least the columns "input_gene" and "ortholog_gene".
}
\description{
Map orthologs from one species to another.
}
\details{
\code{map_orthologs()} is a core function within 
\code{convert_orthologs()}, but does not have many
of the extra checks, such as \code{non121_strategy})
and \code{drop_nonorths}.
}
\examples{
data("exp_mouse")
gene_map <- map_orthologs(genes=rownames(exp_mouse),
                          input_species="mouse")
}
