% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_orthologs.R
\name{map_orthologs}
\alias{map_orthologs}
\title{Map orthologs}
\usage{
map_orthologs(
  gene_df,
  gene_col,
  input_species,
  output_species = "human",
  verbose = TRUE
)
}
\arguments{
\item{gene_df}{Data.frame containing the genes symbols of the input species.}

\item{gene_col}{Character string indicating either the 
input species' gene symbols are in a column with the 
input species gene symbols (\code{gene_col="<gene_col_name>"}) 
or are the row names (\code{gene_col="rownames"}).}

\item{input_species}{Name of the input species (e.g., "mouse","fly"). 
See \code{taxa_id_dict()} for a full list of available species.}

\item{output_species}{Name of the output species (e.g. "human").}

\item{verbose}{Print messages.}
}
\value{
Ortholog map data.frame
}
\description{
Map orthologs from one species to another.
}
\examples{
data("exp_mouse")
gene_df <- data.frame(Gene=rownames(exp_mouse))
gene_map <- map_orthologs(gene_df=gene_df,  gene_col="Gene", input_species="mouse")
}
