% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_species.R
\name{map_species}
\alias{map_species}
\title{Standardise species names}
\usage{
map_species(
  species = NULL,
  search_cols = c("display_name", "id", "scientific_name", "taxonomy_id"),
  output_format = c("id", "display_name", "scientific_name", "taxonomy_id", "version"),
  use_genomeinfodbdata = FALSE,
  use_local = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{species}{Species query
(e.g. "human", "homo sapiens", "hapiens", or 9606).
If given a list, will iterate queries for each item.
Set to \code{NULL} to return all species.}

\item{search_cols}{Which columns to search for
\code{species} substring in
metadata \href{https://biit.cs.ut.ee/gprofiler/api/util/organisms_list}{
API}.}

\item{output_format}{Which column to return.}

\item{use_genomeinfodbdata}{Retrieve an additional 2+ million organisms
from \code{GenomeInfoDb::specData}.
\emph{NOTE: } Not all of these organisms are available
for gene/ortholog mapping.}

\item{use_local}{If \code{TRUE} \emph{default},
\link[orthogene]{map_species}
uses a locally stored version of the species metadata table
instead of pulling directly from the gprofiler API.
Local version may not be fully up to date,
but should suffice for most use cases.}

\item{verbose}{Print messages.}
}
\value{
Species ID of type \code{output_format}
}
\description{
Search \code{gprofiler} database for species
that match the input text string.
Then translate to a standardised species ID.
}
\examples{
ids <- map_species(species = c(
    "human", 9606, "mus musculus",
    "fly", "C elegans"
))
}
