% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_orthologs.R
\name{report_orthologs}
\alias{report_orthologs}
\title{Identify the number of orthologous genes between two species}
\usage{
report_orthologs(
  target_species = "mouse",
  reference_species = "human",
  standardise_genes = FALSE,
  method = c("homologene", "gorth"),
  drop_nonorths = TRUE,
  one_to_one_only = TRUE,
  round_digits = 2,
  return_report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{target_species}{Target species.}

\item{reference_species}{Reference species.}

\item{standardise_genes}{If \code{TRUE} genes will first be standardised 
to HGNC symbols using \code{gprofiler2::gconvert}. 
This will add a new column to \code{gene_df} 
called "input_gene_standard".}

\item{method}{R package to to use for gene mapping: 
\code{"gorth"} (slower but more reference species) or 
\code{"homologene"} (faster but fewer reference species).}

\item{drop_nonorths}{Drop genes that don't have an ortholog 
in the \code{output_species}.}

\item{one_to_one_only}{Drop genes that don't have a 1:1 mapping 
between input species and human 
(i.e. drop genes with multiple human orthologs).}

\item{round_digits}{Number of digits to round to when printing percentages.}

\item{return_report}{Return just the ortholog mapping between two species (\code{FALSE}) 
or return both the ortholog mapping as well a dataframe of the report statistics (\code{TRUE}).}

\item{verbose}{Print messages.}
}
\value{
List of ortholog report statistics
}
\description{
\code{ewceData::mouse_to_human_homologs()} provides 15,604 1:1 mouse:human orthologs.
\code{homologene}-based functions like \code{EWCE::convert_orthologs} 
provide an improved 16,766 1:1 mouse:human orthologs.
}
\examples{
orth.fly <- report_orthologs(target_species = "fly",
                             reference_species="human")
orth.zeb <- report_orthologs(target_species = "zebrafish", 
                             reference_species="human")
orth.mus <- report_orthologs(target_species = "mouse", 
                             reference_species="human") 
}
